pyproject.toml
[tool.poetry]
name = "s3 proxy service"
version = "0.1.0"
description = "Backend Proxy Services for storing and retrieving files in AWS S3."
authors = ["Miraj Tariq"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
boto3 = "^1.35.14"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.6"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
